infrastructure/lambda/data-api-proxies/requirements.txt
boto3>=1.28.0
psycopg2-binary>=2.9.0
opensearch-py>=2.3.0
requests-aws4auth>=1.2.0
